+++ src/management/Kontent.Ai.Management.Tests/Annotations/ContentTypeAndElementAttributesTests.cs
@@ -7,10 +7,10 @@ namespace Kontent.Ai.Management.Tests.Annotations;
 public class ContentTypeAndElementAttributesTests
 {
     [Fact]
-    public void KontentContentType_StoresCodenameAndOptionalId()
+    public void ContentType_StoresCodenameAndOptionalId()
     {
-        var withId = new KontentTypeAttribute("article", "11111111-1111-1111-1111-111111111111");
-        var withoutId = new KontentTypeAttribute("article");
+        var withId = new ContentTypeAttribute("article", "11111111-1111-1111-1111-111111111111");
+        var withoutId = new ContentTypeAttribute("article");
 
         withId.Codename.Should().Be("article");
         withId.Id.Should().Be("11111111-1111-1111-1111-111111111111");
@@ -20,16 +20,16 @@ public class ContentTypeAndElementAttributesTests
     }
 
     [Fact]
-    public void KontentContentType_NullCodename_Throws()
+    public void ContentType_NullCodename_Throws()
     {
-        Action act = () => _ = new KontentTypeAttribute(null!);
+        Action act = () => _ = new ContentTypeAttribute(null!);
         act.Should().Throw<ArgumentNullException>().And.ParamName.Should().Be("codename");
     }
 
     [Fact]
-    public void KontentContentType_UsageTargetsClassOnly()
+    public void ContentType_UsageTargetsClassOnly()
     {
-        var usage = typeof(KontentTypeAttribute).GetCustomAttributes(typeof(AttributeUsageAttribute), false)
+        var usage = typeof(ContentTypeAttribute).GetCustomAttributes(typeof(AttributeUsageAttribute), false)
             .Cast<AttributeUsageAttribute>().Single();
 
         usage.ValidOn.Should().Be(AttributeTargets.Class);
@@ -38,9 +38,9 @@ public class ContentTypeAndElementAttributesTests
     }
 
     [Fact]
-    public void KontentElement_StoresCodenameAndId()
+    public void ContentElement_StoresCodenameAndId()
     {
-        var attr = new KontentElementAttribute("title", "22222222-2222-2222-2222-222222222222");
+        var attr = new ContentElementAttribute("title", "22222222-2222-2222-2222-222222222222");
 
         attr.Codename.Should().Be("title");
         attr.Id.Should().Be("22222222-2222-2222-2222-222222222222");
@@ -49,16 +49,16 @@ public class ContentTypeAndElementAttributesTests
     [Theory]
     [InlineData(null, "id", "codename")]
     [InlineData("codename", null, "id")]
-    public void KontentElement_NullArgs_Throw(string? codename, string? id, string expectedParam)
+    public void ContentElement_NullArgs_Throw(string? codename, string? id, string expectedParam)
     {
-        Action act = () => _ = new KontentElementAttribute(codename!, id!);
+        Action act = () => _ = new ContentElementAttribute(codename!, id!);
         act.Should().Throw<ArgumentNullException>().And.ParamName.Should().Be(expectedParam);
     }
 
     [Fact]
-    public void KontentElement_UsageTargetsPropertyOnly()
+    public void ContentElement_UsageTargetsPropertyOnly()
     {
-        var usage = typeof(KontentElementAttribute).GetCustomAttributes(typeof(AttributeUsageAttribute), false)
+        var usage = typeof(ContentElementAttribute).GetCustomAttributes(typeof(AttributeUsageAttribute), false)
             .Cast<AttributeUsageAttribute>().Single();
 
         usage.ValidOn.Should().Be(AttributeTargets.Property);
@@ -66,9 +66,9 @@ public class ContentTypeAndElementAttributesTests
     }
 
     [Fact]
-    public void KontentEnumValue_StoresCodenameAndId()
+    public void ContentOption_StoresCodenameAndId()
     {
-        var attr = new KontentEnumValueAttribute("news", "33333333-3333-3333-3333-333333333333");
+        var attr = new ContentOptionAttribute("news", "33333333-3333-3333-3333-333333333333");
 
         attr.Codename.Should().Be("news");
         attr.Id.Should().Be("33333333-3333-3333-3333-333333333333");
@@ -77,16 +77,16 @@ public class ContentTypeAndElementAttributesTests
     [Theory]
     [InlineData(null, "id", "codename")]
     [InlineData("codename", null, "id")]
-    public void KontentEnumValue_NullArgs_Throw(string? codename, string? id, string expectedParam)
+    public void ContentOption_NullArgs_Throw(string? codename, string? id, string expectedParam)
     {
-        Action act = () => _ = new KontentEnumValueAttribute(codename!, id!);
+        Action act = () => _ = new ContentOptionAttribute(codename!, id!);
         act.Should().Throw<ArgumentNullException>().And.ParamName.Should().Be(expectedParam);
     }
 
     [Fact]
-    public void KontentEnumValue_UsageTargetsFieldOnly()
+    public void ContentOption_UsageTargetsFieldOnly()
     {
-        var usage = typeof(KontentEnumValueAttribute).GetCustomAttributes(typeof(AttributeUsageAttribute), false)
+        var usage = typeof(ContentOptionAttribute).GetCustomAttributes(typeof(AttributeUsageAttribute), false)
             .Cast<AttributeUsageAttribute>().Single();
 
         usage.ValidOn.Should().Be(AttributeTargets.Field);
@@ -94,31 +94,31 @@ public class ContentTypeAndElementAttributesTests
     }
 
     // End-to-end: applying the attributes to a stub generated record reads back the same values via reflection.
-    [KontentType("article", "44444444-4444-4444-4444-444444444444")]
+    [ContentType("article", "44444444-4444-4444-4444-444444444444")]
     private sealed class StubArticle
     {
-        [KontentElement("title", "55555555-5555-5555-5555-555555555555")]
+        [ContentElement("title", "55555555-5555-5555-5555-555555555555")]
         public string? Title { get; init; }
     }
 
     private enum StubCategory
     {
-        [KontentEnumValue("news", "66666666-6666-6666-6666-666666666666")] News,
+        [ContentOption("news", "66666666-6666-6666-6666-666666666666")] News,
     }
 
     [Fact]
     public void Attributes_RoundTripThroughReflection()
     {
-        var type = typeof(StubArticle).GetCustomAttribute<KontentTypeAttribute>()!;
+        var type = typeof(StubArticle).GetCustomAttribute<ContentTypeAttribute>()!;
         type.Codename.Should().Be("article");
 
         var prop = typeof(StubArticle).GetProperty(nameof(StubArticle.Title))!
-            .GetCustomAttribute<KontentElementAttribute>()!;
+            .GetCustomAttribute<ContentElementAttribute>()!;
         prop.Codename.Should().Be("title");
         prop.Id.Should().Be("55555555-5555-5555-5555-555555555555");
 
-        var enumValue = typeof(StubCategory).GetField(nameof(StubCategory.News))!
-            .GetCustomAttribute<KontentEnumValueAttribute>()!;
-        enumValue.Codename.Should().Be("news");
+        var option = typeof(StubCategory).GetField(nameof(StubCategory.News))!
+            .GetCustomAttribute<ContentOptionAttribute>()!;
+        option.Codename.Should().Be("news");
     }
 }